Upgrade addresses to 64 bits (fixed size)

+++ src/libc/uart/uart.c
@@ -2,15 +2,15 @@
 #include "../misc.h"
 
 // Memory-Mapped I/O output
-static inline void mmio_write(uint32_t reg, uint32_t data)
+static inline void mmio_write(uint64_t reg, uint64_t data)
 {
-	*(volatile uint32_t*)reg = data;
+	*(volatile uint64_t *) reg = data;
 }
 
 // Memory-Mapped I/O input
-static inline uint32_t mmio_read(uint32_t reg)
+static inline uint64_t mmio_read(uint64_t reg)
 {
-	return *(volatile uint32_t*)reg;
+	return *(volatile uint64_t *)reg;
 }
 
 void uart_init()
